// ambari-web/latest/src/screens/Services/highAvailibility/nameNode/store/types.ts
export interface State {
  enableHighAvailibilitySteps: any;
}

export enum ActionTypes {
  STORE_INFORMATION = "STORE INFORMATION",
  SYNC_STATE = "SYNC STATE",
  REMOVE_KEY = "REMOVE KEY",
}

export type Action =
  | { type: ActionTypes.STORE_INFORMATION; payload: any }
  | { type: ActionTypes.SYNC_STATE; payload: any }
  | { type: ActionTypes.REMOVE_KEY; payload: any };